data_collection.py
```python
# Import necessary libraries
import requests
from bs4 import BeautifulSoup
import pandas as pd
import numpy as np
import cv2
from PIL import Image
import os

class DataCollector:
    def __init__(self):
        self.data_dir = 'medical_data/'

    def collect_patient_records(self, source_url):
        # This is a placeholder function. In a real-world scenario, you would need to access a secure and private database
        # to collect patient records. The implementation of this function will depend on the specifics of such a database.
        pass

    def collect_diagnostic_images(self, source_url):
        # This is a placeholder function. In a real-world scenario, you would need to access a secure and private database
        # to collect diagnostic images. The implementation of this function will depend on the specifics of such a database.
        pass

    def collect_genetic_info(self, source_url):
        # This is a placeholder function. In a real-world scenario, you would need to access a secure and private database
        # to collect genetic information. The implementation of this function will depend on the specifics of such a database.
        pass

    def collect_clinical_notes(self, source_url):
        # This is a placeholder function. In a real-world scenario, you would need to access a secure and private database
        # to collect clinical notes. The implementation of this function will depend on the specifics of such a database.
        pass

    def preprocess_data(self):
        # This is a placeholder function. In a real-world scenario, you would need to preprocess the collected data
        # to make it suitable for further analysis. The implementation of this function will depend on the specifics of the data.
        pass

if __name__ == "__main__":
    collector = DataCollector()
    collector.collect_patient_records('source_url')
    collector.collect_diagnostic_images('source_url')
    collector.collect_genetic_info('source_url')
    collector.collect_clinical_notes('source_url')
    collector.preprocess_data()
```
