requirements.txt
# Quantum Computing
qiskit==0.23.6

# Machine Learning and Data Analysis
numpy==1.20.1
pandas==1.2.3
scikit-learn==0.24.1
tensorflow==2.4.1
keras==2.4.3

# Image Processing
opencv-python==4.5.1.48
Pillow==8.1.2

# Data Collection
beautifulsoup4==4.9.3
requests==2.25.1

# Security
cryptography==3.4.7

# User Interface
flask==1.1.2

# Documentation
sphinx==3.5.3

# Ethical Considerations
aif360==0.3.0

# Patent Application
uspto-patent-client==0.6.2
